Finish restructing rss module
